Refactor components and summaryComponents and added spesific TS type configuration
