+++ src/layout/TabGroup/SummaryTabGroupComponent.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+import type { SummaryRendererProps } from 'src/layout/LayoutComponent';
+
+// FIXME: Implement SummaryTabGroupComponent
+
+export const SummaryTabGroupComponent = (props: SummaryRendererProps<'TabGroup'>) => <div>Summary</div>;
